feat: removed circular dependencies between ml and data
