+++ phymes-core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod metrics;
+pub mod schemas;
 pub mod session;
 pub mod table;
 pub mod task;
