+++ phymes-core/src/schemas/chat_completion.rs
