+++ phymes-core/src/schemas/embedding.rs
