+++ phymes-core/src/schemas/mod.rs
@@ -1,0 +1,7 @@
+pub mod message_history;
+
+// Based on openai-api-rs <https://github.com/dongri/openai-api-rs>
+pub mod chat_completion;
+pub mod common;
+pub mod embedding;
+pub mod types;
