+++ phymes-ml/src/candle_assets/mod.rs
@@ -1,4 +1,3 @@
 pub mod candle_asset;
 pub mod candle_which;
-pub mod device;
 pub mod token_output_stream;
