+++ phymes-ml/src/candle_chat/mod.rs
@@ -1,6 +1,5 @@
 pub mod chat_config;
 pub mod chat_processor;
 pub mod message_aggregator_processor;
-pub mod message_history;
 pub mod message_parser_processor;
 pub mod tool_parser;
