+++ phymes-ml/src/openai_asset/mod.rs
@@ -17,9 +17,3 @@ pub enum OpenAIRequestState {
 }
 
 pub mod openai_which;
-
-// Based on openai-api-rs <https://github.com/dongri/openai-api-rs>
-pub mod chat_completion;
-pub mod common;
-pub mod embedding;
-pub mod types;
